EDI - Allow getting excel-files from remote sftp
- application.properties: change xls.order.* to excel.order.* to make it clearer
- also rename class files from XLS* to Excel*
- ExcelOLCandRoute: introduce optional remote-to-local-route

https://github.com/metasfresh/metasfresh/issues/6801
