+++ misc/services/edi/src/main/java/de/metas/edi/esb/bean/imports/excel/ExcelImpCOLCandTypeBuilder.java
@@ -1,10 +1,8 @@
-package de.metas.edi.esb.jaxb;
-
 /*
  * #%L
- * de.metas.edi.esb
+ * de-metas-edi-esb-camel
  * %%
- * Copyright (C) 2015 metas GmbH
+ * Copyright (C) 2020 metas GmbH
  * %%
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as
@@ -13,23 +11,22 @@ package de.metas.edi.esb.jaxb;
  *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public
- * License along with this program.  If not, see
+ * License along with this program. If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
  * #L%
  */
 
+package de.metas.edi.esb.bean.imports.excel;
 
 import static de.metas.edi.esb.commons.Util.resolveGenericLookup;
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
 
-import de.metas.edi.esb.bean.imports.xls.XLSConfigurationContext;
-import de.metas.edi.esb.bean.imports.xls.XLS_OLCand_Row;
 import de.metas.edi.esb.commons.Constants;
 import de.metas.edi.esb.commons.Util;
 import de.metas.edi.esb.jaxb.metasfresh.COrderDeliveryRuleEnum;
@@ -48,16 +45,16 @@ import de.metas.edi.esb.jaxb.metasfresh.XLSImpCOLCandType;
  * @author tsa
  * @task 08839
  */
-public class XLSImpCOLCandTypeBuilder
+public class ExcelImpCOLCandTypeBuilder
 {
-	public static final XLSImpCOLCandTypeBuilder newBuilder()
+	public static final ExcelImpCOLCandTypeBuilder newBuilder()
 	{
-		return new XLSImpCOLCandTypeBuilder();
+		return new ExcelImpCOLCandTypeBuilder();
 	}
 
 	private final XLSImpCOLCandType olcand;
 
-	private XLSImpCOLCandTypeBuilder()
+	private ExcelImpCOLCandTypeBuilder()
 	{
 		super();
 
@@ -75,7 +72,7 @@ public class XLSImpCOLCandTypeBuilder
 		return olcand;
 	}
 
-	public XLSImpCOLCandTypeBuilder setFromContext(final XLSConfigurationContext ctx)
+	public ExcelImpCOLCandTypeBuilder setFromContext(final ExcelConfigurationContext ctx)
 	{
 		// set ReplicationTrx attribute
 		olcand.setTrxNameAttr(ctx.getCamelFileName());
@@ -108,7 +105,7 @@ public class XLSImpCOLCandTypeBuilder
 		return this;
 	}
 
-	public XLSImpCOLCandTypeBuilder setFromRow(final XLS_OLCand_Row row)
+	public ExcelImpCOLCandTypeBuilder setFromRow(final Excel_OLCand_Row row)
 	{
 		olcand.setLine(toBigIntegerOrNull(row.getLineNo()));
 		olcand.setPOReference(row.getPOReference());
