+++ misc/services/edi/src/main/resources/application.properties
@@ -11,9 +11,12 @@ server.port=8184
 
 #
 # SFTP settings
+# privateKeyFiles shall be placed into {{edi.local.dir}}/ssh/ when using SFTP
 edi.sftpserver.userHostAndPort=editeststorage.metasfresh.com:22
-# to be placed into {{edi.local.dir}}/ssh/ when using SFTP
 edi.sftpserver.privateKeyFile=mySshPrivateKeyFileName
+
+excel.sftserver.userHostAndPort=editeststorage.metasfresh.com:22
+excel.sftserver.privateKeyFile=mySshPrivateKeyFileName
 
 #
 #Endpoint URLs for imported and exported EDI files
@@ -47,7 +50,7 @@ edi.file.invoic.stepcom-xml=file://{{edi.local.dir}}/{{edi.file.invoic.subdir}}/
 #rabbitmq config
 #note: when running within docker, our startup-script overrides these 4 properties
 camel.component.rabbitmq.hostname=localhost
-camel.component.rabbitmq.port-number=30050
+camel.component.rabbitmq.port-number=5672
 camel.component.rabbitmq.username=guest
 camel.component.rabbitmq.password=guest
 
@@ -210,12 +213,15 @@ edi.props.invoic.isStepComXML=true
 #
 # Excel Orders Import (task 08839)
 # Code reference: XLSConfigurationContext
-xls.order.input=file://{{edi.local.dir}}/{{edi.file.orders.subdir}}/mf-excel-pricelist
-xls.order.currency=CHF
-xls.order.ad_client_value={{edi.order.ad_client_value}}
-xls.order.ad_org_id={{edi.order.ad_org_id}}
-xls.order.ad_input_datadestination_internalname={{edi.order.ad_input_datadestination_internalname}}
-xls.order.ad_input_datasource_id=160
-xls.order.ad_user_enteredby_id=160
-xls.order.deliveryrule=A
-xls.order.deliveryviarule=P
+
+excel.file.orders=file://{{edi.local.dir}}/{{edi.file.orders.subdir}}/mf-excel-pricelist?preMove=processessing&move=../done&moveFailed=../error
+excel.file.orders.remote=sftp://{{excel.sftpserver.userHostAndPort}}/{{edi.file.orders.subdir}}?privateKeyFile={{edi.local.dir}}/ssh/{{excel.sftpserver.privateKeyFile}}
+
+excel.order.currency=CHF
+excel.order.ad_client_value={{edi.order.ad_client_value}}
+excel.order.ad_org_id={{edi.order.ad_org_id}}
+excel.order.ad_input_datadestination_internalname={{edi.order.ad_input_datadestination_internalname}}
+excel.order.ad_input_datasource_id=160
+excel.order.ad_user_enteredby_id=160
+excel.order.deliveryrule=A
+excel.order.deliveryviarule=P
